Fix release method: no version in js file banner
Bower doesn't support a version property in its manifest (`bower.json`) so I
removed it so the focus is on npm and its versioning system. But I forgot to
change the loading of bower.json to package.json to get the version number for
the distribution files' banners (file headers).