Test correct angular version is installed
Look at the `bower.json` file of all the Angular libs we are using and test if
the `version` property matches the given version, which is set in a matrix in
`.travis.yml`.

Always log to the console instead of only when DEBUG=1

This uses bin files so a `node -e 'require()'` kind of script isn't required to
run it from the shell.